requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.0.0
requests>=2.26.0
numpy>=1.21.0
python-dotenv>=0.19.0
flask>=2.0.0 